Styled the left side of the page.
The left side is the 'hero' section which has a
background image, the company name, & the source
of the background image. I have positioned the text
and applied color & font styles. I also linked a new
stylesheet for the font I am using in this section.
Ready to move onto the next section.